# perf: Optimize pandas `GeoArrowExtensionArray.copy()`