--- a/geoarrow-pandas/tests/test_geoarrow_pandas.py
+++ b/geoarrow-pandas/tests/test_geoarrow_pandas.py
@@ -113,6 +113,40 @@ def test_array_basic_methods():
         ),
     )
 
+def test_array_basic_methods_chunked_data():
+    pa_array = ga.array(["POINT (0 1)", "POINT (1 2)", None])
+    array = gapd.GeoArrowExtensionArray(pa.chunked_array([pa_array]))
+
+    assert array[0] == gapd.GeoArrowExtensionScalar("POINT (0 1)")
+    assert array[2] is None
+    assert isinstance(array[1:2], gapd.GeoArrowExtensionArray)
+    assert len(array[1:2]) == 1
+    assert array[1:2][0] == gapd.GeoArrowExtensionScalar("POINT (1 2)")
+    assert isinstance(array[[1]], gapd.GeoArrowExtensionArray)
+    assert array[[1]][0] == gapd.GeoArrowExtensionScalar("POINT (1 2)")
+
+    assert len(array) == 3
+    assert all(array[:2] == array[:2])
+    assert array.dtype == gapd.GeoArrowExtensionDtype(ga.wkt())
+    assert array.nbytes == pa_array.nbytes
+    assert isinstance(array.take(np.array([1])), gapd.GeoArrowExtensionArray)
+    assert array.take(np.array([1]))[0] == gapd.GeoArrowExtensionScalar("POINT (1 2)")
+    np.testing.assert_array_equal(array.isna(), np.array([False, False, True]))
+
+    assert isinstance(array.copy(), gapd.GeoArrowExtensionArray)
+    assert array.copy()[0] == gapd.GeoArrowExtensionScalar("POINT (0 1)")
+
+    np.testing.assert_array_equal(
+        array.to_numpy(),
+        np.array(
+            [
+                gapd.GeoArrowExtensionScalar("POINT (0 1)"),
+                gapd.GeoArrowExtensionScalar("POINT (1 2)"),
+                None,
+            ]
+        ),
+    )
+
 
 def test_array_concat():
     pa_array_wkt = ga.array(["POINT (0 1)", "POINT (1 2)", None])