Credit to @martinfleis for the epic traceback that highlighted this and @jorisvandenbossche for the `concat_arrays()` trick!

```python
import geoarrow.pandas as gapd
import geoarrow.pyarrow as ga
import numpy as np

pts_array = ga.point().from_geobuffers(
    None, np.random.random(int(1e6)), np.random.random(int(1e6))
)

pts_pd = gapd.GeoArrowExtensionArray(pts_array)

%timeit pts_pd.copy()
#> Before this PR:
#> 37.8 s ± 188 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
#> After this PR:
#> 630 µs ± 22.4 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```