@kylebarron 2023-10-05
```
%timeit pts_pd.copy()
#> Before this PR:
#> 37.8 s ± 188 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
#> After this PR:
#> 630 µs ± 22.4 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```
that's a good speedup 😉